clean duplicate code for async job
