+++ broker/src/integration-test/groovy/com/swisscom/cloud/sb/broker/util/DummyConfig.groovy
@@ -1,9 +1,10 @@
 package com.swisscom.cloud.sb.broker.util
 
+import com.swisscom.cloud.sb.broker.services.AsyncServiceConfig
 import com.swisscom.cloud.sb.broker.services.bosh.BoshBasedServiceConfig
 import groovy.transform.CompileStatic
 
 
 @CompileStatic
-class DummyConfig implements BoshBasedServiceConfig {
+class DummyConfig implements BoshBasedServiceConfig, AsyncServiceConfig {
 }
