+++ broker/src/main/groovy/com/swisscom/cloud/sb/broker/services/AsyncServiceConfig.groovy
@@ -1,0 +1,11 @@
+package com.swisscom.cloud.sb.broker.services
+
+import com.swisscom.cloud.sb.broker.cfextensions.endpoint.EndpointConfig
+import com.swisscom.cloud.sb.broker.services.bosh.BoshConfig
+import groovy.transform.CompileStatic
+
+@CompileStatic
+trait AsyncServiceConfig implements EndpointConfig, BoshConfig {
+    int retryIntervalInSeconds
+    int maxRetryDurationInMinutes
+}
