+++ broker/src/main/groovy/com/swisscom/cloud/sb/broker/services/AsyncServiceProvider.groovy
@@ -1,9 +1,11 @@
-package com.swisscom.cloud.sb.broker.services.kubernetes.service
+package com.swisscom.cloud.sb.broker.services
 
 import com.swisscom.cloud.sb.broker.async.AsyncProvisioningService
 import com.swisscom.cloud.sb.broker.cfextensions.endpoint.EndpointLookup
+import com.swisscom.cloud.sb.broker.cfextensions.endpoint.EndpointProvider
 import com.swisscom.cloud.sb.broker.model.DeprovisionRequest
 import com.swisscom.cloud.sb.broker.model.ProvisionRequest
+import com.swisscom.cloud.sb.broker.model.ServiceInstance
 import com.swisscom.cloud.sb.broker.provisioning.DeprovisionResponse
 import com.swisscom.cloud.sb.broker.provisioning.ProvisionResponse
 import com.swisscom.cloud.sb.broker.provisioning.ProvisioningPersistenceService
@@ -15,14 +17,14 @@ import com.swisscom.cloud.sb.broker.provisioning.job.ServiceDeprovisioningJob
 import com.swisscom.cloud.sb.broker.provisioning.job.ServiceProvisioningJob
 import com.swisscom.cloud.sb.broker.services.common.ServiceProvider
 import com.swisscom.cloud.sb.broker.services.common.Utils
-import com.swisscom.cloud.sb.broker.services.kubernetes.config.KubernetesConfig
+import com.swisscom.cloud.sb.model.endpoint.Endpoint
 import groovy.transform.CompileStatic
 import groovy.util.logging.Slf4j
 import org.springframework.beans.factory.annotation.Autowired
 
 @CompileStatic
 @Slf4j
-abstract class KubernetesAsyncServiceProvider<T extends KubernetesConfig> implements ServiceProvider, AsyncServiceProvisioner, AsyncServiceDeprovisioner {
+abstract class AsyncServiceProvider<T extends AsyncServiceConfig> implements ServiceProvider, AsyncServiceProvisioner, AsyncServiceDeprovisioner, EndpointProvider {
 
     @Autowired
     protected AsyncProvisioningService asyncProvisioningService
@@ -49,6 +51,11 @@ abstract class KubernetesAsyncServiceProvider<T extends KubernetesConfig> implem
         asyncProvisioningService.scheduleDeprovision(new DeprovisioningJobConfig(ServiceDeprovisioningJob.class, request,
                 serviceConfig.retryIntervalInSeconds, serviceConfig.maxRetryDurationInMinutes))
         return new DeprovisionResponse(isAsync: true)
+    }
+
+    @Override
+    Collection<Endpoint> findEndpoints(ServiceInstance serviceInstance) {
+        return endpointLookup.findEndpoints(serviceInstance, serviceConfig)
     }
 
 }
