+++ broker/src/main/groovy/com/swisscom/cloud/sb/broker/services/kubernetes/config/KubernetesConfig.groovy
@@ -2,6 +2,7 @@ package com.swisscom.cloud.sb.broker.services.kubernetes.config
 
 import com.swisscom.cloud.sb.broker.cfextensions.endpoint.EndpointConfig
 import com.swisscom.cloud.sb.broker.config.Config
+import com.swisscom.cloud.sb.broker.services.AsyncServiceConfig
 import groovy.transform.AutoClone
 import groovy.transform.CompileStatic
 import org.springframework.boot.context.properties.ConfigurationProperties
@@ -11,12 +12,9 @@ import org.springframework.context.annotation.Configuration
 @CompileStatic
 @Configuration
 @ConfigurationProperties(prefix = 'com.swisscom.cloud.sb.broker.service.kubernetes')
-class KubernetesConfig implements Config, EndpointConfig {
+class KubernetesConfig implements Config, EndpointConfig, AsyncServiceConfig {
     String kubernetesHost
     String kubernetesPort
     String kubernetesClientPFXPath
     String kubernetesClientPFXPasswordPath
-    int retryIntervalInSeconds
-    int maxRetryDurationInMinutes
-
 }
