+++ broker/src/test/groovy/com/swisscom/cloud/sb/broker/services/bosh/DummyConfig.groovy
@@ -1,5 +1,7 @@
 package com.swisscom.cloud.sb.broker.services.bosh
 
+import com.swisscom.cloud.sb.broker.services.AsyncServiceConfig
 
-class DummyConfig implements BoshBasedServiceConfig {
+
+class DummyConfig implements BoshBasedServiceConfig, AsyncServiceConfig {
 }
